summer_clip/trainers_utils/requirements.txt
omegaconf
wandb
ptflops
pytorch_gan_metrics